xia: separate general FIB from list FIB
This change creates separate files for a FIB that is implemented
as a list of buckets from the general code for FIBs. This change
makes it easier to add other FIB implementations.
