+++ net/xia/Makefile
@@ -4,8 +4,8 @@
 
 obj-$(CONFIG_XIA) += xia.o
 
-xia-objs :=	main.o socket.o dag.o locktbl.o fib.o fib_frontend.o route.o \
-		output.o vxidty.o
+xia-objs :=	main.o socket.o dag.o locktbl.o list_fib.o fib.o \
+		fib_frontend.o route.o output.o vxidty.o
 
 obj-$(CONFIG_XIA_PPAL_AD) += ppal_ad/
 obj-$(CONFIG_XIA_PPAL_HID) += ppal_hid/
