+++ net/xia/ppal_ad/main.c
@@ -1,5 +1,5 @@
 #include <linux/module.h>
-#include <net/xia_fib.h>
+#include <net/xia_list_fib.h>
 #include <net/xia_route.h>
 #include <net/xia_dag.h>
 #include <net/xia_vxidty.h>
