+++ net/xia/ppal_hid/main.c
@@ -1,5 +1,6 @@
 #include <linux/module.h>
 #include <net/xia_dag.h>
+#include <net/xia_list_fib.h>
 #include <net/xia_output.h>
 #include <net/xia_vxidty.h>
 #include <net/xia_hid.h>
