+++ net/xia/ppal_hid/nwp.c
@@ -5,6 +5,7 @@
 #include <linux/cache.h>
 #include <net/xia_dag.h>
 #include <net/xia_hid.h>
+#include <net/xia_list_fib.h>
 
 /* Neighbor Table */
